added variant prop to Footer, created styles different Footer variants, created FooterNotLoggedIn and FooterRegistration components
